requirements.txt
PyQt6>=6.7,<7.0
PyQt6-Qt6>=6.7,<7.0
pyserial>=3.5,<4.0
